requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.6.4
python-dotenv==1.2.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
google-generativeai==0.8.3
